server/requirements.txt
Flask==3.1.1
google-genai==1.30.0
requests==2.31.0
python-dotenv==1.0.0